telegram-bot/requirements.txt
python-telegram-bot==20.7
aiohttp==3.9.1
pydantic==2.5.0
structlog==23.2.0
tenacity==8.2.3
python-dotenv==1.0.0
loguru==0.7.2 
aiogram